requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
ultralytics==8.3.141
easyocr==1.7.0
pillow==10.1.0
python-multipart==0.0.6
numpy==1.24.3
opencv-python-headless==4.8.1.78
torch>=2.0.0,<2.2.0
torchvision>=0.15.0,<0.17.0
psutil==5.9.6